:card_file_box: adjusting code folder structure
