+++ plotly_resampler/figure_resampler/figure_resampler_interface.py
@@ -1,11 +1,6 @@
 # -*- coding: utf-8 -*-
 """
-Wrapper around the plotly ``go.Figure`` class which allows bookkeeping and
-back-end based resampling of high-frequency sequential data.
-
-Tip
----
-The term `high-frequency` actually refers very large amounts of sequential data.
+Abstract interface of which the concrete *Resampler* classes write towards.
 
 """
 
@@ -14,7 +9,6 @@ from __future__ import annotations
 __author__ = "Jonas Van Der Donckt, Jeroen Van Der Donckt, Emiel Deprost"
 
 import re
-import warnings
 from copy import copy
 from typing import Dict, Iterable, List, Optional, Tuple, Union
 from uuid import uuid4
@@ -23,13 +17,10 @@ import dash
 import numpy as np
 import pandas as pd
 import plotly.graph_objects as go
-from jupyter_dash import JupyterDash
-from dash import Dash
 from plotly.basedatatypes import BaseTraceType, BaseFigure
-from trace_updater import TraceUpdater
 
-from .aggregation import AbstractSeriesAggregator, EfficientLTTB
-from .utils import round_td_str, round_number_str
+from ..aggregation import AbstractSeriesAggregator, EfficientLTTB
+from ..utils import round_td_str, round_number_str
 
 from abc import ABC
 
@@ -950,316 +941,3 @@ class AbstractFigureAggregator(BaseFigure, ABC):
             if m is not None:
                 matches.append(m.string)
         return sorted(matches)
-
-
-class _FigureWidgetResamplerM(type(AbstractFigureAggregator), type(go.FigureWidget)):
-    # MetaClass for the FigureWidgetResampler
-    pass
-
-
-class FigureWidgetResampler(
-    AbstractFigureAggregator, go.FigureWidget, metaclass=_FigureWidgetResamplerM
-):
-    """Data aggregation functionality wrapper for ``go.FigureWidgets``."""
-
-    def __init__(
-        self,
-        figure: go.FigureWidget,
-        convert_existing_traces: bool = True,
-        default_n_shown_samples: int = 1000,
-        default_downsampler: AbstractSeriesAggregator = EfficientLTTB(),
-        resampled_trace_prefix_suffix: Tuple[str, str] = (
-            '<b style="color:sandybrown">[R]</b> ',
-            "",
-        ),
-        show_mean_aggregation_size: bool = True,
-        verbose: bool = False,
-    ):
-        if not isinstance(figure, go.FigureWidget):
-            figure = go.FigureWidget(figure)
-
-        super().__init__(
-            figure,
-            convert_existing_traces,
-            default_n_shown_samples,
-            default_downsampler,
-            resampled_trace_prefix_suffix,
-            show_mean_aggregation_size,
-            verbose,
-        )
-
-        self._prev_layout = None  # Contains the previous xaxis layout configuration
-
-        # used for logging purposes to save a history of layout changes
-        self._relayout_hist = []
-
-        # A list of al xaxis string names e.g., "xaxis", "xaxis2", "xaxis3", ....
-        self._xaxis_list = self._re_matches(re.compile("xaxis\d*"), self._layout.keys())
-        # edge case: an empty `go.Figure()` does not yet contain xaxis keys
-        if not len(self._xaxis_list):
-            self._xaxis_list = ['xaxis']
-
-        # Assign the the update-methods to the corresponding classes
-        showspike_keys = [f"{xaxis}.showspikes" for xaxis in self._xaxis_list]
-        self.layout.on_change(self._update_spike_ranges, *showspike_keys)
-
-        x_relayout_keys = [f"{xaxis}.range" for xaxis in self._xaxis_list]
-        self.layout.on_change(self._update_x_ranges, *x_relayout_keys)
-
-    def _update_x_ranges(self, layout, *x_ranges):
-        """Update the the go.Figure data based on changed x-ranges.
-
-        Parameters
-        ----------
-        layout : go.Layout
-            The figure's (i.e, self) layout object. Remark that this is a reference,
-            so if we change self.layout (same object reference), this object will change.
-        *x_ranges: iterable
-            A iterable list of current x-ranges, where each x-range is a tuple of two
-            items, indicating the current/new (if changed) left-right x-range,
-            respectively.
-        """
-        relayout_dict = {}  # variable in which we aim to reconstruct the relayout
-        # serialize the layout in a new dict object
-        layout = {
-            xaxis_str: layout[xaxis_str].to_plotly_json()
-            for xaxis_str in self._xaxis_list
-        }
-        if self._prev_layout is None:
-            self._prev_layout = layout
-
-        for xaxis_str, x_range in zip(self._xaxis_list, x_ranges):
-            # We also check whether "range" is within the xaxis its layout otherwise
-            # It is most-likely an autorange check
-            if (
-                "range" in layout[xaxis_str]
-                and self._prev_layout[xaxis_str].get("range", []) != x_range
-            ):
-                # a change took place -> add to the relayout dict
-                relayout_dict[f"{xaxis_str}.range[0]"] = x_range[0]
-                relayout_dict[f"{xaxis_str}.range[1]"] = x_range[1]
-
-                # An update will take place for that trace 
-                # -> save current xaxis range to _prev_layout
-                self._prev_layout[xaxis_str]['range'] = x_range
-
-        if len(relayout_dict):
-            # Construct the update data
-            update_data = self.construct_update_data(relayout_dict)
-
-            if self._print_verbose:
-                self._relayout_hist.append(dict(zip(self._xaxis_list, x_ranges)))
-                self._relayout_hist.append(layout)
-                self._relayout_hist.append(["xaxis-range-update", len(update_data) - 1])
-                self._relayout_hist.append("-" * 30)
-
-            with self.batch_update():
-                # First update the layout (first item of update_data)
-                self.layout.update(update_data[0])
-
-                for xaxis_str in self._xaxis_list:
-                    if 'showspikes' in layout[xaxis_str]:
-                        self.layout[xaxis_str].pop("showspikes")
-
-                # Then update the data
-                for updated_trace in update_data[1:]:
-                    trace_idx = updated_trace.pop("index")
-                    self.data[trace_idx].update(updated_trace)
-
-    def _update_spike_ranges(self, layout, *showspikes):
-        """Update the go.Figure based on the changed spike-ranges.
-
-        Parameters
-        ----------
-        layout : go.Layout
-            The figure's (i.e, self) layout object. Remark that this is a reference,
-            so if we change self.layout (same object reference), this object will change.
-        *showspikes: iterable
-            A iterable where each item is a bool, indicating  whether showspikes is set
-            to true/false for the corresponding xaxis in ``self._xaxis_list``.
-        """
-        relayout_dict = {}  # variable in which we aim to reconstruct the relayout
-        # serialize the layout in a new dict object
-        layout = {
-            xaxis_str: layout[xaxis_str].to_plotly_json()
-            for xaxis_str in self._xaxis_list
-        }
-
-        if self._prev_layout is None:
-            return
-
-        for xaxis_str, showspike in zip(self._xaxis_list, showspikes):
-            if (
-                # autorange key must be set to True
-                layout[xaxis_str].get("autorange", False)
-                # we only perform updates for traces which have 'range' property,
-                # as we do need to reconstruct the update-data for these traces
-                and self._prev_layout[xaxis_str].get("range", None) is not None
-            ):
-                relayout_dict[f"{xaxis_str}.autorange"] = True
-                relayout_dict[f"{xaxis_str}.showspikes"] = showspike
-                # autorange -> we pop the xaxis range
-                if 'range' in layout[xaxis_str]:
-                    del layout[xaxis_str]['range']
-
-        if len(relayout_dict):
-            # An update will take place, save current layout to _prev_layout
-            self._prev_layout = layout
-
-            # Construct the update data
-            update_data = self.construct_update_data(relayout_dict)
-            if self._print_verbose:
-                self._relayout_hist.append(layout)
-                self._relayout_hist.append(["showspikes-update", len(update_data) - 1])
-                self._relayout_hist.append("-" * 30)
-
-            with self.batch_update():
-                # First update the layout (first item of update_data)
-                self.layout.update(update_data[0])
-
-                # Also:  Remove the showspikes from the layout, otherwise the autorange
-                # will not work as intended (it will not be triggered again)
-                # Note: this removal causes a second trigger of this method
-                # which will go in the "else" part below.
-                for xaxis_str in self._xaxis_list:
-                    self.layout[xaxis_str].pop("showspikes")
-
-                # Then, update the data
-                for updated_trace in update_data[1:]:
-                    trace_idx = updated_trace.pop("index")
-                    self.data[trace_idx].update(updated_trace)
-        elif self._print_verbose:
-            self._relayout_hist.append(["showspikes", "initial call or showspikes"])
-            self._relayout_hist.append("-" * 40)
-
-    # def show(self, *args, **kwargs):
-    #     super(go.FigureWidget, self).show(*args, **kwargs)
-
-
-class FigureResampler(AbstractFigureAggregator, go.Figure):
-    """Data aggregation functionality for Figures."""
-
-    def __init__(
-        self,
-        figure: go.Figure = go.Figure(),
-        convert_existing_traces: bool = True,
-        default_n_shown_samples: int = 1000,
-        default_downsampler: AbstractSeriesAggregator = EfficientLTTB(),
-        resampled_trace_prefix_suffix: Tuple[str, str] = (
-            '<b style="color:sandybrown">[R]</b> ',
-            "",
-        ),
-        show_mean_aggregation_size: bool = True,
-        verbose: bool = False,
-    ):
-        assert isinstance(figure, go.Figure)
-        super().__init__(
-            figure,
-            convert_existing_traces,
-            default_n_shown_samples,
-            default_downsampler,
-            resampled_trace_prefix_suffix,
-            show_mean_aggregation_size,
-            verbose,
-        )
-
-        # The figureAggregator needs a dash app
-        self._app: JupyterDash | Dash | None = None
-        self._port: int | None = None
-        self._host: str | None = None
-
-    def show_dash(
-        self,
-        mode=None,
-        config: dict | None = None,
-        graph_properties: dict | None = None,
-        **kwargs,
-    ):
-        """Registers the :func:`update_graph` callback & show the figure in a dash app.
-
-        Parameters
-        ----------
-        mode: str, optional
-            Display mode. One of:\n
-              * ``"external"``: The URL of the app will be displayed in the notebook
-                output cell. Clicking this URL will open the app in the default
-                web browser.
-              * ``"inline"``: The app will be displayed inline in the notebook output
-                cell in an iframe.
-              * ``"jupyterlab"``: The app will be displayed in a dedicated tab in the
-                JupyterLab interface. Requires JupyterLab and the ``jupyterlab-dash``
-                extension.
-            By default None, which will result in the same behavior as ``"external"``.
-        config: dict, optional
-            The configuration options for displaying this figure, by default None.
-            This ``config`` parameter is the same as the dict that you would pass as
-            ``config`` argument to the `show` method.
-            See more https://plotly.com/python/configuration-options/
-        graph_properties: dict, optional
-            Dictionary of (keyword, value) for the properties that should be passed to
-            the dcc.Graph, by default None.
-            e.g.: {"style": {"width": "50%"}}
-            Note: "config" is not allowed as key in this dict, as there is a distinct
-            ``config`` parameter for this property in this method.
-            See more https://dash.plotly.com/dash-core-components/graph
-        **kwargs: dict
-            Additional app.run_server() kwargs.
-            e.g.: port
-
-        """
-        graph_properties = {} if graph_properties is None else graph_properties
-        assert "config" not in graph_properties.keys()  # There is a param for config
-        # 1. Construct the Dash app layout
-        app = JupyterDash("local_app")  # TODO -> was jupyterdash
-        app.layout = dash.html.Div(
-            [
-                dash.dcc.Graph(
-                    id="resample-figure", figure=self, config=config, **graph_properties
-                ),
-                TraceUpdater(
-                    id="trace-updater", gdID="resample-figure", sequentialUpdate=False
-                ),
-            ]
-        )
-        self.register_update_graph_callback(app, "resample-figure", "trace-updater")
-
-        # 2. Run the app
-        if (
-            self.layout.height is not None
-            and mode == "inline"
-            and "height" not in kwargs
-        ):
-            # If figure height is specified -> re-use is for inline dash app height
-            kwargs["height"] = self.layout.height + 18
-
-        # store the app information, so it can be killed
-        self._app = app
-        self._host = kwargs.get("host", "127.0.0.1")
-        self._port = kwargs.get("port", "8050")
-
-        app.run_server(mode=mode, **kwargs)
-
-    def stop_server(self, warn: bool = True):
-        """Stop the running dash-app.
-
-        Parameters
-        ----------
-        warn: bool
-            Whether a warning message will be shown or  not, by default True.
-
-        .. attention::
-            This only works if the dash-app was started with :func:`show_dash`.
-        """
-        if self._app is not None:
-
-            old_server = self._app._server_threads.get((self._host, self._port))
-            if old_server:
-                old_server.kill()
-                old_server.join()
-                del self._app._server_threads[(self._host, self._port)]
-        elif warn:
-            warnings.warn(
-                "Could not stop the server, either the \n"
-                + "\t- 'show-dash' method was not called, or \n"
-                + "\t- the dash-server wasn't started with 'show_dash'"
-            )
